Analyzed CSVs, file system cleanup, api key removed
